Add Schema::hasTable to migrations

+++ database/migrations/2022_07_04_121543_create_boiler_water_table.php
@@ -13,15 +13,17 @@ class CreateBoilerWaterTable extends Migration
      */
     public function up()
     {
-        Schema::create('boiler_water', function (Blueprint $table) {
-            $table->increments('id');
-            $table->float('set_value')->nullable()->default(NULL);
-            $table->float('min_value')->nullable()->default(NULL);
-            $table->float('max_value')->nullable()->default(NULL);
+        if (!Schema::hasTable('boiler_water')) {
+            Schema::create('boiler_water', function (Blueprint $table) {
+                $table->increments('id');
+                $table->float('set_value')->nullable()->default(NULL);
+                $table->float('min_value')->nullable()->default(NULL);
+                $table->float('max_value')->nullable()->default(NULL);
 
-            $table->foreign('id_object')->references('id')->on('objects')
-                  ->onUpdate('cascade')->onDelete('set null');
-        });
+                $table->foreign('id_object')->references('id')->on('objects')
+                      ->onUpdate('cascade')->onDelete('set null');
+            });
+        }
     }
 
     /**
